requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.3
matplotlib>=3.8.2
seaborn>=0.13.1
plotly>=5.18.0
PyPDF2>=3.0.1
python-docx>=0.8.11
anthropic>=0.5.0
openai>=1.69.0
requests>=2.31.0
pillow>=10.0.0
python-dotenv>=1.0.0
matplotlib
numpy
pandas
plotly
seaborn
streamlit
openpyxl
python-docx
openpyxl
pdfkit==1.0.0